Docs: Fixed type inference example in Numerics section
This is a fix to a passage in Numeric types. There are two examples of type inference here (e.g. let x = 42; // x has type i32). However, the examples are introduced with the following statement: "If a number literal has nothing to cause its type to be inferred, it defaults:" These are actual examples of type inference, so it doesn't make sense to say that there is "nothing to cause its type to be inferred" only to then demonstrate its type being inferred... Changed it to: "If the type is not specified, the compiler infers the type based on the assigned value:" (I also wanted to be explicit about who/what is making the inference – the compiler).